% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockDiagonal.R
\name{block_diagonal}
\alias{block_diagonal}
\title{Given list of matrices, form block sparse matrix}
\usage{
block_diagonal(mats)
}
\arguments{
\item{mats}{List of sparse matrices \code{dgCMatrix}.}
}
\value{
A block diagonal \code{dgCMatrix} with each element of mats
along the diagonal.
}
\description{
Given list of matrices, form block sparse matrix
}
