% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_influence_func.R
\name{get_adjs}
\alias{get_adjs}
\title{Returns small-sample adjustment}
\usage{
get_adjs(est, cluster = NULL)
}
\arguments{
\item{est}{Estimation object of class \code{fixest}
and \code{fixest_multi} from the \code{fixest} package}

\item{cluster}{Scalar string denoting the name
of the cluster column.}
}
\value{
A vector of degrees of freedoms
}
\description{
Returns small-sample adjustment
}
