% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_influence_func.R
\name{get_cluster}
\alias{get_cluster}
\title{Returns (X'X)^{-1} X'e from regression}
\usage{
get_cluster(est, cluster, use_rowid)
}
\arguments{
\item{est}{Estimation object of class \code{fixest}
and \code{fixest_multi} from the \code{fixest} package}

\item{cluster}{Scalar string denoting the name
of the cluster column.}

\item{use_rowid}{Scalar logical denoting if row number
should be used as the cluster.}
}
\value{
A vector of cluster labels
}
\description{
Returns (X'X)^{-1} X'e from regression
}
