% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_influence_func.R
\name{get_dof}
\alias{get_dof}
\title{Returns degrees of freedom from regression}
\usage{
get_dof(est)
}
\arguments{
\item{est}{Estimation object of class \code{fixest}
and \code{fixest_multi} from the \code{fixest} package}
}
\value{
A vector of degrees of freedoms
}
\description{
Returns degrees of freedom from regression
}
