% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_influence_func.R
\name{get_influence_func}
\alias{get_influence_func}
\title{Returns (X'X)^{-1} X'e from regression}
\usage{
get_influence_func(est)
}
\arguments{
\item{est}{Estimation object of class \code{lm} or
\code{fixest} and \code{fixest_multi} from the \code{fixest} package}
}
\value{
A matrix with rows consisting of (X'X)^{-1} X_i'e_i.
If \code{fixest_multi} is provided, then returns a
block diagonal matrix of influence functions.
}
\description{
Returns (X'X)^{-1} X'e from regression
}
