% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_hypotheses.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_coef}
\alias{set_coef}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{marginaleffects}{\code{\link[marginaleffects]{get_coef}}, \code{\link[marginaleffects]{set_coef}}}
}}

