% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fixest_list.R
\name{split_fixest_list}
\alias{split_fixest_list}
\title{Split fixest_multi into seperate elements of a list}
\usage{
split_fixest_list(ests)
}
\description{
Split fixest_multi into seperate elements of a list
}
