% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sur_hypotheses.R
\name{sur_hypotheses}
\alias{sur_hypotheses}
\title{Hypothesis testing from seemingly-unrelated regressions}
\usage{
sur_hypotheses(
  ests,
  cluster = NULL,
  hypothesis = NULL,
  conf_level = 0.95,
  df = NULL,
  joint = FALSE,
  joint_test = "f"
)
}
\arguments{
\item{ests}{List of \code{feols} estimates.}

\item{cluster}{Scalar string denoting the name of the cluster column. If
the same dataset is used (or subsets across samples) and the cluster
variable "rowid" is passed, this will use the row-number of the original
data. This with \code{ssadj = TRUE} will obtain the original standard errors.}

\item{hypothesis}{String containing the hypothesis to test.
See \code{marginaleffects::hypotheses} for more details. Additionally,
see the description for more details to specify which regression
a coefficient is from in the hypothesis.}

\item{conf_level}{numeric value between 0 and 1. Confidence level to use to build a confidence interval.}

\item{df}{Degrees of freedom used to compute p values and confidence intervals. A single numeric value between 1 and \code{Inf}. When \code{df} is \code{Inf}, the normal distribution is used. When \code{df} is finite, the \code{t} distribution is used. See \link[insight:get_df]{insight::get_df} for a convenient function to extract degrees of freedom. Ex: \code{slopes(model, df = insight::get_df(model))}}

\item{joint}{Joint test of statistical significance. The null hypothesis value can be set using the \code{hypothesis} argument.
\itemize{
\item FALSE: Hypotheses are not tested jointly.
\item TRUE: All parameters are tested jointly.
\item String: A regular expression to match parameters to be tested jointly. \code{grep(joint, perl = TRUE)}
\item Character vector of parameter names to be tested. Characters refer to the names of the vector returned by \code{coef(object)}.
\item Integer vector of indices. Which parameters positions to test jointly.
}}

\item{joint_test}{A character string specifying the type of test, either "f" or "chisq". The null hypothesis is set by the \code{hypothesis} argument, with default null equal to 0 for all parameters.}
}
\description{
This function takes a set of regression esimates and
perform hypothesis tests using \code{vcovSUR} to calculate the
variance-covariance matrix between regressions. The underlying
hypothesis test is conducted using \code{marginaleffects::hypotheses}.
Since the same regressors can be used in multiple regressions, the
coefficients will be suffixed with \verb{_#} with the \verb{#} corresponding to
the estimates position in the \code{ests} object.
}
