% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSUR.R
\name{vcovSUR}
\alias{vcovSUR}
\title{Get joint variance-covariance matrix of seemingly unrelated regressions}
\usage{
vcovSUR(ests, cluster = NULL)
}
\arguments{
\item{ests}{List of \code{feols} estimates.}

\item{cluster}{Scalar string denoting the name of the cluster column.}
}
\value{
Variance-covariance matrix of the combined set of coefficients in
all models in \code{ests}.
}
\description{
Get joint variance-covariance matrix of seemingly unrelated regressions
}
