% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovSUR.R
\name{vcovSUR}
\alias{vcovSUR}
\title{Get joint variance-covariance matrix of seemingly unrelated regressions}
\usage{
vcovSUR(ests, cluster = NULL, ssadj = FALSE)
}
\arguments{
\item{ests}{List of \code{feols} estimates.}

\item{cluster}{Scalar string denoting the name of the cluster column. If
the same dataset is used (or subsets across samples) and the cluster
variable "rowid" is passed, this will use the row-number of the original
data. This with \code{ssadj = TRUE} will obtain the original standard errors.}

\item{ssadj}{Logical indicating whether to apply a small-sample adjustment.
See details for how the adjustment is made.}
}
\value{
Variance-covariance matrix of the combined set of coefficients in
all models in \code{ests}.
}
\description{
Get joint variance-covariance matrix of seemingly unrelated regressions
}
\details{
The small-sample adjustment depends on the \code{cluster} option and the \code{ssadj}
option. For each estimate, the degree of freedom calculation will be
calculated. Denote them as c_i. For block V_ij, the adjustment will
be sqrt(c_i * c_j) so that the diagonal elements are adjusted by c_i.
This matches \code{dfk} option in \code{sureg}. If \code{cluster} is used (and is not
"rowid"), then the adjustment will be G/(G-1) where G is the
number of unique clusters. If \code{cluster} is \code{NULL} or is \code{"rowid"}, then
the adjustment will be (N_i - 1)(N_i - K_i) where N_i is the
number of observations and K_i is the number of coefficients used in
estimate i.

If a \code{fixest_multi} object is used and \code{cluster} is not specified, then
\code{cluster} will automatically be set to \code{"rowid"} since the same dataset
is used in all estimates.
}
